// src/models/Data.ts
export type Data = {
  id: number;
  first_name: string;
  last_name: string;
  email: string;
  phone_number: string;
  address_street: string;
  address_city: string;
  address_state: string;
  address_zip: string;
  address_country: string;
  address_latitude: number;
  address_longitude: number;
  birth_date: string;
  birth_place: string;
  birth_country: string;
  company: string;
  job_title: string;
  registration_date: string;
  registration_ip: string;
  registration_city: string;
  registration_state: string;
  registration_zip: string;
  registration_country: string;
  registration_latitude: number;
  registration_longitude: number;
  account_balance: number;
  account_number: string;
  account_type: string;
  account_currency: string;
  account_status: string;
  account_created_at: string;
  account_updated_at: string;
  account_last_login: string;
  wishlist_items_count: number;
  wishlist_items: string[];
  cart_items_count: number;
  cart_items: string[];
  orders_count: number;
  orders: string[];
  reviews_count: number;
  reviews: string[];
  browser: string;
  os: string;
  device: string;
  ip_address: string;
  user_agent: string;
  created_at: string;
  updated_at: string;
  is_active: boolean;
  credit_card_last4: string;
  credit_card_brand: string;
  credit_card_exp_month: number;
  credit_card_exp_year: number;
  credit_card_type: string;
  credit_card_issuer: string;
  credit_card_country: string;
  credit_card_city: string;
  gender: string;
  age: number;
  height: number;
  weight: number;
  blood_type: string;
  eye_color: string;
  hair_color: string;
  hair_length: string;
  login_attempts: number;
  login_attempts_failed: number;
  login_attempts_success: number;
  login_attempts_total: number;
  login_attempts_last_attempt: string;
  login_attempts_last_attempt_success: boolean;
  login_attempts_last_attempt_failed: boolean;
  login_attempts_last_attempt_total: number;
  medical_conditions: string[];
  allergies: string[];
  medications: string[];
  medical_history: string[];
  medical_history_last_update: string;
  medical_history_last_update_by: string;
  medical_history_last_update_by_id: number;
  medical_history_last_update_by_name: string;
  time_zone_offset: number;
  time_zone_name: string;
  time_zone_abbreviation: string;
  time_zone_current_time: string;
  time_zone_current_time_unix: number;
  time_zone_current_time_unix_ms: number;
  time_zone_current_time_unix_ns: number;
  time_zone_current_time_unix_us: number;
  locale: string;
  currency: string;
  language: string;
  timezone: string;
  last_address_change: string;
  last_address_change_ip: string;
  last_address_change_user_agent: string;
  last_address_change_created_at: string;
  last_address_change_updated_at: string;
  last_address_change_is_active: boolean;
  last_address_change_credit_card_last4: string;
  orer_list: [
    {
      id: number;
      order_number: string;
      order_date: string;
      order_total: number;
      order_status: string;
    },
  ];
};